# optimize performance of paginate()