@arj03 2020-12-14
Overall the implementation is small and easy to understand. Just want to make sure that we don't end with nasty surprises later on :)

> heapUsed grows about +23MB every page, and seems to be garbage collected every 2 pages or so.

This means that if you do 10 different pagination, the memory will not go above around 50mb? I'm guessing these numbers come from stress testing this doing a lot of queries to see if you can cause some kind of OOM situation?