@staltz 2020-12-14
> I'm guessing these numbers come from stress testing this doing a lot of queries to see if you can cause some kind of OOM situation?

Did you see the before and after in the memory section? The numbers come from the same benchmark code, I just added a `console.log(process.memoryUsage())` in the pull drain. The memory in the `before` case grows because for each page, it creates a new bitset and then creates a new sorted timestamp array, and these are quite big when there are ~200k items.