@staltz 2020-12-14
@arj03 I removed the `log.since`, I didn't benchmark with `migrate` because the new solution is lazy, not eager, I can't see how it would affect migrate now.