@staltz 2020-12-14
Check out this before and after, getting 1000 pages worth of posts:

## Before

<details>
<summary>Click for animated GIF</summary>

![before](https://user-images.githubusercontent.com/90512/102072981-4948c400-3e0b-11eb-8a76-f8bd9cba0a45.gif)

</details>

## After

<details>
<summary>Click for animated GIF</summary>

![after](https://user-images.githubusercontent.com/90512/102072984-4c43b480-3e0b-11eb-9827-d6207dfa9880.gif)

</details>