@arj03 2020-12-14
Much more comfortable with this now without the log.since. Really nice use of WeakMap.